Fix method return capture in the presence of ret branches
In some methods returns weren't captured correctly. This happened
when there was a branch straight to the return; it kept pointing
at the return after instrumentation.

This change fixes this, and also turns on single return
instrumentation (I haven't seen any issues with it with this change,
fingers crosses); error messages that would have been commonly
caused by this problem were changed to warn level.