Capture generic object values correctly
Codegen for Object.ToString() call was too simplistic. It failed if the
module didn't have an existing reference to System.Object or Object.ToString.
Also it didn't use constrained prefix to callvirt correctly, thus breaking
in some common cases, such as Nullable<T>. Now the references are emitted
if missing, and the value is correctly stored in a local, address taken
and a constrained virtual call to ToString generated.